scripts/swarm/python-runner.py
import sys
import os
import traceback

def main():
    if len(sys.argv) < 2:
        print("Missing instruction payload", file=sys.stderr)
        sys.exit(1)
        
    instruction = sys.argv[1]
    
    # 1. Initialize API Client
    api_key = os.environ.get("GEMINI_API_KEY")
    if not api_key:
        print("Missing GEMINI_API_KEY", file=sys.stderr)
        sys.exit(1)
        
    print(f"Starting Python agent execution for instruction: {instruction}")
    
    try:
        from google import genai
        from google.genai import types
        
        client = genai.Client(api_key=api_key)
        
        # 2. System Instruction
        system_instruction = (
            "You are an autonomous Python AI swarm worker inside an isolated repository worktree. "
            "You have access to bash commands, file reading, and file writing. "
            "Execute your assigned instruction and write your output to the required location. "
            "When you are completely finished, output the exact word 'DONE'."
        )
        
        # 3. Dummy iteration loop for the agent (mirroring TS runner)
        # Note: In a complete implementation, you'd define ADK tool functions 
        # for 'read_file', 'write_file', 'exec_command' exactly as in TS.
        
        print("AGENT: Processing...", flush=True)
        # Placeholder for actual task execution logic...
        
        print("DONE", flush=True)
        sys.exit(0)
        
    except ImportError as e:
        print(f"ImportError: {e}. Please ensure google-genai is installed.", file=sys.stderr)
        sys.exit(1)
    except Exception as e:
        print(f"Error in python agent loop: {e}", file=sys.stderr)
        traceback.print_exc()
        sys.exit(1)

if __name__ == "__main__":
    main()
