scripts/swarm/requirements.txt
google-genai
# google-adk (or any other required ADK packages)
